+++ source/constants.h
@@ -2,11 +2,11 @@
 
 #include <limits>
 
-__device__ constexpr double kInfinity = std::numeric_limits<double>::infinity();
-__device__ constexpr double kPi = 3.1415926535897932385;
+__device__ constexpr float kInfinity = std::numeric_limits<float>::infinity();
+__device__ constexpr float kPi = 3.1415926535897932385;
 
 // Utility Functions
 
-inline double degrees_to_radians(double degrees) {
-    return degrees * kPi / 180.0;
+inline float degrees_to_radians(float degrees) {
+    return degrees * kPi / 180.0f;
 }
