+++ source/hittable_list.h
@@ -8,7 +8,7 @@ struct HittableList : public Hittable {
     __host__ __device__  HittableList() {}
     __host__ __device__  HittableList(Hittable **list, int size): m_list(list), m_size(size) {}
 
-    __host__ __device__ bool Hit(const Ray& ray, double ray_tmin, double ray_tmax, HitRecord& rec) const override {
+    __host__ __device__ bool Hit(const Ray& ray, float ray_tmin, float ray_tmax, HitRecord& rec) const override {
         HitRecord temp_rec;
         bool hit_anything = false;
         auto closest_so_far = ray_tmax;
