+++ source/vec3.h
@@ -7,15 +7,15 @@ template <typename T>
 class Vec3Base {
   public:
     __host__ __device__ Vec3Base() {}
-    __host__ __device__ Vec3Base(double e0, double e1, double e2) : e{e0, e1, e2} {}
+    __host__ __device__ Vec3Base(float e0, float e1, float e2) : e{e0, e1, e2} {}
 
-    __host__ __device__ double X() const { return e[0]; }
-    __host__ __device__ double Y() const { return e[1]; }
-    __host__ __device__ double Z() const { return e[2]; }
+    __host__ __device__ float X() const { return e[0]; }
+    __host__ __device__ float Y() const { return e[1]; }
+    __host__ __device__ float Z() const { return e[2]; }
 
     __host__ __device__ Vec3Base<T> operator-() const { return Vec3Base<T>(-e[0], -e[1], -e[2]); }
-    __host__ __device__ double operator[](int i) const { return e[i]; }
-    __host__ __device__ double& operator[](int i) { return e[i]; }
+    __host__ __device__ float operator[](int i) const { return e[i]; }
+    __host__ __device__ float& operator[](int i) { return e[i]; }
 
     __host__ __device__ Vec3Base<T>& operator+=(const Vec3Base<T>& v) {
         e[0] += v.e[0];
@@ -24,14 +24,14 @@ class Vec3Base {
         return *this;
     }
 
-    __host__ __device__ Vec3Base<T>& operator*=(double t) {
+    __host__ __device__ Vec3Base<T>& operator*=(float t) {
         e[0] *= t;
         e[1] *= t;
         e[2] *= t;
         return *this;
     }
 
-    __host__ __device__ Vec3Base<T>& operator/=(double t) {
+    __host__ __device__ Vec3Base<T>& operator/=(float t) {
         return *this *= 1/t;
     }
 
@@ -40,16 +40,16 @@ class Vec3Base {
         return T(e[0], e[1], e[2]);
     }
 
-    __host__ __device__ double Length() const {
+    __host__ __device__ float Length() const {
         return sqrt(LengthSquared());
     }
 
-    __host__ __device__ double LengthSquared() const {
+    __host__ __device__ float LengthSquared() const {
         return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
     }
 
   protected:
-    double e[3];
+    float e[3];
 };
 
 using Vec3 = Vec3Base<void>;
@@ -77,22 +77,22 @@ __host__ __device__ Vec3Base<T> operator*(const Vec3Base<T>& u, const Vec3Base<T
 }
 
 template <typename T>
-__host__ __device__ Vec3Base<T> operator*(double t, const Vec3Base<T>& v) {
+__host__ __device__ Vec3Base<T> operator*(float t, const Vec3Base<T>& v) {
     return Vec3Base<T>(t*v[0], t*v[1], t*v[2]);
 }
 
 template <typename T>
-__host__ __device__ Vec3Base<T> operator*(const Vec3Base<T>& v, double t) {
+__host__ __device__ Vec3Base<T> operator*(const Vec3Base<T>& v, float t) {
     return t * v;
 }
 
 template <typename T>
-__host__ __device__ Vec3Base<T> operator/(const Vec3Base<T>& v, double t) {
+__host__ __device__ Vec3Base<T> operator/(const Vec3Base<T>& v, float t) {
     return (1/t) * v;
 }
 
 template <typename T>
-__host__ __device__ double Dot(const Vec3Base<T>& u, const Vec3Base<T>& v) {
+__host__ __device__ float Dot(const Vec3Base<T>& u, const Vec3Base<T>& v) {
     return u[0] * v[0]
          + u[1] * v[1]
          + u[2] * v[2];
